metodos especificos de usuario foram removidos do controlador da API, foi arrumada classe DO do usuario que estava trazendo informações erradas, metodo para realizar login por api criado e validado
